swipe-from-left opens drawer only if current tab is at top of navigation stack
Otherwise, swipe-from-left should go back in the navigation stack.

To accomplish this we add “isAtTop” state to RootView, and set up a
callback prop “onNavigate” in NavigatorWithBar to change that state
when a route is pushed to the navigation stack.

To make that happen, we provide a new “navigate” function in context
that calls the callback instead of having child components call
navigate.push directly.