persist entire state to AsyncStorage after every change
This is the “simplest thing that could possibly work” implementation —
we may run into concurrency issues, performance issues, etc. in the
future but for now this does the job just fine thank you very much, and
requires very little new thinking. (the main thing to remember is to
avoid creating a new state object in the reducer if the value is the
same as it was; see e.g. the change to CHECK_SESSION handling.)

i’m leaving a console.log in with timing info so we can keep an eye on
performance when we start firing lots of actions.